refactor(ui): move renderSyncAttempts + lifecycle into diagnostics/, barrel the tab
Finish the diagnostics.ts decomposition. renderSyncAttempts and
renderSyncDiagnosticsUnavailable move into
diagnostics/render/sync-attempts.ts. renderPairing, renderPairingCollapsible,
renderRedactControl, setRenderSyncPeers, and initDiagnosticsEvents move
into diagnostics/lifecycle.ts (keeps the cross-module render callback
state on module-scope closures so the redact toggle can reach
renderSyncPeers). diagnostics.ts becomes a thin barrel re-exporting
the six public entrypoints.
